add nearby router
