+++ src/index.js
@@ -5,6 +5,7 @@ import { isLoggedOn, addMiddlewares } from './middlewares';
 
 const express = require('express');
 const passport = require('passport');
+const nearbyRoute = require('./routes/nearby');
 
 require('dotenv').config();
 
@@ -30,6 +31,8 @@ app.use('/newUser', newUserRouter);
 
 app.use('/createUser', newUserRouter);
 
+app.use('/', nearbyRoute);
+
 app.get('/login', function (req, res) {
   res.render('login');
 });
@@ -49,29 +52,6 @@ app.get('/logout', function (req, res) {
 
 app.get('/profile', isLoggedOn, function (req, res) {
   res.render('profile', { user: req.user });
-});
-
-app.get('/nearby', function (req, res) {
-  let gear = {
-    temp: 50,
-    elevationGain: 2005,
-    boots: 'https://www.flaticon.com/svg/static/icons/svg/2826/2826618.svg',
-    lightJacket:
-      'https://www.flaticon.com/svg/static/icons/svg/3126/3126039.svg',
-    heavyJacket: 'https://www.flaticon.com/svg/static/icons/svg/614/614247.svg',
-    poles: 'https://www.flaticon.com/svg/static/icons/svg/2325/2325148.svg',
-    desBoots:
-      'Wear a solid pair of hiking boots as the terrain can be challenging',
-    desHeavyJacket: "Bring a heavier jacket, it's below 32°F",
-    desLightJacket: "Bring a lighter jacket, it's above 32°F",
-    desPoles: 'Bring hiking poles, the elevation gain is more than 2000 feet',
-  };
-
-  res.render('nearby', { gear: gear });
-});
-
-app.get('/gear', function (req, res) {
-  res.render('gear');
 });
 
 // start server
